// generated/models/User.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type User = {
    account?: string;
    avatar?: string;
    birth?: string;
    create_time?: string;
    email?: string;
    gender?: string;
    hobby?: string;
    is_delete?: number;
    password?: string;
    phone?: string;
    profile?: string;
    rating?: number;
    readings?: number;
    role?: number;
    school?: string;
    tags?: string;
    update_time?: string;
    url?: string;
    username?: string;
    uuid?: number;
    background_picture?:string
};